requirements.txt
openai
langchain
streamlit
psycopg2-binary
python-dotenv
